Refactor dictionary construction
