+++ compiler/lib/src/main/scala/analysis/Semantics/ConstructDictionary/DictionaryUsedSymbols.scala
@@ -2,7 +2,7 @@ package fpp.compiler.analysis
 
 import fpp.compiler.util._
 
-/** Computes the used symbols for a dictionary */
+/** Fills in the used symbols for a dictionary */
 case class DictionaryUsedSymbols(a: Analysis, t: Topology) {
 
   def updateUsedSymbols(d: Dictionary): Dictionary =
